requirements.txt
Flask==2.1.0
flask-cors==3.0.10
mysql-connector-python==8.0.26
python-dotenv==0.19.1
